# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel blog Api Documentation",
        "description": "You will get all blog api",
        "contact": {
            "email": "ankush0094@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/laravel_blog_api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/v1/category": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Category listing",
                "description": "get all category",
                "operationId": "Category",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        },
                        "example": "10"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/category/add": {
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Add Category",
                "description": "Add new category",
                "operationId": "Add Category",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "slug"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "slug": {
                                        "type": "string"
                                    },
                                    "parent_id": {
                                        "type": "integer"
                                    },
                                    "media": {
                                        "type": "file"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "demo name",
                                    "slug": "demo slug"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/category/edit/{id}": {
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "updated category",
                "description": "Update category",
                "operationId": "Update category",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "category id parameter",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "slug"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "slug": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "demo title",
                                    "content": "content title",
                                    "tags[0]": "tag1",
                                    "tags[1]": "tag2"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/category/delete/{id}": {
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Delete category",
                "description": "Delete category",
                "operationId": "Delete category",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "category id parameter",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "422": {
                        "description": "validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/post": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Post listing",
                "description": "get all post",
                "operationId": "Listing",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        },
                        "example": "20"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}